Improved datastore handling in Get-UnProtectedVm
As per https://github.com/benmeadowcroft/SRM-Cmdlets/issues/12#issuecomment-56302137 there was a bug in how we searched for VMs on a protected datastore. I've updated the code to hopefully address this as suggested by nilanshakya. I expect to have an ABR environment back up and running in a week or two so hopefully I can test this more fully then and remove the TODO marker.