Get-UnProtectedVM to only show VMs with .vmx files on protected datastores
If a VM has the .vmx file stored on a non-protected datastore, but has VMDK files located on protected datastores, they do not show up in the SRM GUI, but Get-UnProtectedVM function will show it.  Since they cannot be resolved by Protect-VM, this update filters only VMs that have the .VMX file located on the protected datastores.